Created login and has applied component
