+++ apps/financial-aid/web-osk/graphql/sharedGql.ts
@@ -57,6 +57,7 @@ export const GetCurrentUserQuery = gql`
   query currentUserQuery {
     currentUser {
       name
+      hasAppliedForPeriod
     }
   }
 `
