+++ apps/financial-aid/web-osk/src/components/HasApplied/hasApplied.treat.ts
@@ -1,0 +1,2 @@
+import { style } from 'treat'
+import { theme } from '@island.is/island-ui/theme'
